refactor: split runtime command services and status contract
