+++ desktop/src-tauri/src/runtime/mod.rs
@@ -1,8 +1,12 @@
 pub(crate) mod diagnostics;
+pub(crate) mod model_discovery;
 pub(crate) mod operation;
 pub(crate) mod profile;
+pub(crate) mod profile_switch;
 pub(crate) mod provider;
 pub(crate) mod proxy;
+pub(crate) mod proxy_lifecycle;
+pub(crate) mod sandbox_session;
 pub(crate) mod science;
 pub(crate) mod system;
 pub(crate) mod transaction;
