+++ desktop/src-tauri/src/runtime/sandbox_session.rs
@@ -1,0 +1,168 @@
+use std::process::{Command, Stdio};
+use std::time::Duration;
+
+use serde_json::{json, Value};
+
+use crate::runtime::operation::{
+    self, OperationKind, OperationStage, OperationTrace, POLL_INTERVAL_MS,
+};
+use crate::runtime::proxy::ProxyAction;
+use crate::runtime::proxy_lifecycle::ensure_proxy;
+use crate::runtime::science::{sandbox_home, sandbox_running_ours, sandbox_url, stop_sandbox};
+use crate::runtime::system::{asset_root, log_path, open_in_browser, open_log, redact, tail_file};
+use crate::{config, lifecycle, lock, oauth_forge, proc, AppState, SharedAppState};
+
+fn stop_sandbox_state(app: &tauri::AppHandle, st: &mut AppState) -> Result<(), String> {
+    stop_sandbox(app, &mut st.sandbox, &mut st.sandbox_url)
+}
+
+/// One-click session startup: active proxy, virtual login, sandbox, browser.
+///
+/// Callers must hold the command serializer lock.
+pub(crate) fn one_click_login(
+    app: tauri::AppHandle,
+    state: SharedAppState,
+    lifecycle: &lifecycle::Lifecycle,
+) -> Result<Value, String> {
+    let trace = OperationTrace::start(OperationKind::OneClickLogin, "command=one_click_login");
+    let (pport, secret, proxy_action) = ensure_proxy(&app, &state, lifecycle, Some(&trace))?;
+
+    let dir = config::default_dir();
+    let cfg = config::load_from(&dir).map_err(|e| e.to_string())?;
+    let sport = cfg.sandbox_port;
+
+    let sbx_home = sandbox_home();
+    let auth_dir = sbx_home.join(".claude-science");
+
+    if sandbox_running_ours(sport) {
+        if oauth_forge::login_intact(&auth_dir, "virtual@localhost.invalid", &sbx_home) {
+            let url = sandbox_url(sport);
+            {
+                let mut st = lock(&state);
+                st.sandbox_port = sport;
+                st.sandbox_url = Some(url.clone());
+            }
+            let base = match proxy_action {
+                ProxyAction::Reused => "已在运行",
+                ProxyAction::Restarted => "已用新配置重启代理，Science 沿用不变",
+            };
+            let msg = match open_in_browser(&url) {
+                Ok(()) => format!("{base}，已重新打开 Science。"),
+                Err(_) => format!("{base}，服务已就绪，请手动打开：{url}"),
+            };
+            trace.finish(format!(
+                "ok action=reopened proxy_action={}",
+                proxy_action.as_str()
+            ));
+            return Ok(json!({ "url": url, "msg": msg, "action": "reopened" }));
+        }
+        {
+            let mut st = lock(&state);
+            let _ = stop_sandbox_state(&app, &mut st);
+        }
+    }
+
+    let root = asset_root(&app)
+        .ok_or("找不到 scripts/launch-virtual-sandbox.sh（打包资源或仓库根均未命中）。")?;
+
+    trace.stage(OperationStage::SandboxLogin, "ensure_virtual_login");
+    let (forged, login_action) =
+        oauth_forge::ensure_virtual_login(&auth_dir, "virtual@localhost.invalid", &sbx_home)
+            .map_err(|e| format!("写虚拟登录失败：{e}"))?;
+
+    let launch = root.join("scripts/launch-virtual-sandbox.sh");
+    if !launch.is_file() {
+        return Err("找不到 scripts/launch-virtual-sandbox.sh。".into());
+    }
+
+    let proxy_url = format!("http://127.0.0.1:{pport}/{secret}");
+    let logf = open_log("sandbox.log").map_err(|e| format!("建日志失败：{e}"))?;
+    {
+        use std::io::Write;
+        let mut lw = &logf;
+        let _ = writeln!(
+            lw,
+            "[oauth] 虚拟登录已就绪（Rust，零 node；action={:?}）：auth_dir={} account={} org={} enc={}",
+            login_action,
+            forged.auth_dir.display(),
+            forged.account_uuid,
+            forged.org_uuid,
+            forged.enc_file.display()
+        );
+    }
+    let logf2 = logf.try_clone().map_err(|e| e.to_string())?;
+    trace.stage(OperationStage::SandboxLaunch, format!("port={sport}"));
+    let status = Command::new("zsh")
+        .arg(&launch)
+        .arg("--port")
+        .arg(sport.to_string())
+        .arg("--proxy-url")
+        .arg(&proxy_url)
+        .arg("--skip-oauth-forge")
+        .env("SANDBOX_HOME", sandbox_home())
+        .stdout(Stdio::from(logf))
+        .stderr(Stdio::from(logf2))
+        .status()
+        .map_err(|e| format!("起沙箱失败：{e}"))?;
+    if !status.success() {
+        let tail = redact(&tail_file(&log_path("sandbox.log"), 600), &secret);
+        trace.finish("error=sandbox_launch_failed");
+        return Err(format!("起沙箱脚本失败。\n{tail}"));
+    }
+
+    let mut ok = false;
+    for _ in 0..(operation::SANDBOX_HEALTH_BUDGET_MS / POLL_INTERVAL_MS) {
+        std::thread::sleep(Duration::from_millis(POLL_INTERVAL_MS));
+        if proc::http_health(sport, None, operation::LOCAL_HEALTH_TIMEOUT_MS) {
+            ok = true;
+            break;
+        }
+    }
+    trace.stage(
+        OperationStage::SandboxHealth,
+        if ok { "ready" } else { "not_ready" },
+    );
+    if !ok {
+        let tail = redact(&tail_file(&log_path("sandbox.log"), 600), &secret);
+        {
+            let mut st = lock(&state);
+            let _ = stop_sandbox_state(&app, &mut st);
+        }
+        trace.finish("error=sandbox_health_timeout");
+        return Err(format!(
+            "沙箱起后探活超时（端口 {sport}）。已尝试停掉刚起的沙箱。\n{tail}"
+        ));
+    }
+
+    if !sandbox_running_ours(sport) {
+        {
+            let mut st = lock(&state);
+            let _ = stop_sandbox_state(&app, &mut st);
+        }
+        trace.finish("error=sandbox_identity_mismatch");
+        return Err(format!(
+            "端口 {sport} 有服务响应，但按 data-dir 确认不是本沙箱 Science（疑似被其它服务占用）。已尝试停掉刚起的沙箱。"
+        ));
+    }
+
+    let url = sandbox_url(sport);
+    {
+        let mut st = lock(&state);
+        st.sandbox_port = sport;
+        st.sandbox_url = Some(url.clone());
+    }
+    let started = match login_action {
+        oauth_forge::LoginAction::Created => "已启动",
+        _ => "沙箱已重新启动，沿用原有对话",
+    };
+    let msg = match open_in_browser(&url) {
+        Ok(()) => format!("{started}。"),
+        Err(_) => format!("{started}，服务已就绪，请手动打开：{url}"),
+    };
+    trace.stage(OperationStage::OpenBrowser, "done");
+    trace.finish(format!(
+        "ok action=started proxy_action={}",
+        proxy_action.as_str()
+    ));
+    Ok(json!({ "url": url, "msg": msg, "action": "started" }))
+}
